+++ packages/fiori-app-sub-generator/test/unit/utils/common.test.ts
@@ -8,7 +8,7 @@ import { DatasourceType, OdataVersion } from '@sap-ux/odata-service-inquirer';
 import memFs from 'mem-fs';
 import type { Editor } from 'mem-fs-editor';
 import memFsEditor from 'mem-fs-editor';
-import { join } from 'path';
+import { join } from 'node:path';
 import { FloorplanFE, FloorplanFF } from '../../../src/types';
 import { ApiHubType, SapSystemSourceType, minUi5VersionForPageBuildingBlock } from '../../../src/types/constants';
 import {