+++ packages/environment-check/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @sap-ux/environment-check
 
+## 0.18.35
+
+### Patch Changes
+
+-   43a2446: chore: fix Sonar issues
+-   Updated dependencies [43a2446]
+    -   @sap-ux/axios-extension@1.22.10
+    -   @sap-ux/project-access@1.32.3
+    -   @sap-ux/ui5-config@0.29.7
+    -   @sap-ux/btp-utils@1.1.4
+    -   @sap-ux/store@1.1.5
+
 ## 0.18.34
 
 ### Patch Changes